feat: add base classes entrypoints
